fix: run as root, remove priority class
Bootstrap added a non-root user by default, but we need to run as root
in order to create bind mounts. This changes the manifests to run as
root in order to keep compatability with the generated Dockerfile.

Removed a priority class since it's an outreach specific one and we
don't need it in GKE.

Added documentation on how the service works, how to use it, and how to
deploy it.

[DT-3802]